--- app/src/main/java/org/seonhwan/android/veloginmobile/LoginActivity.kt
@@ -1,27 +1,0 @@
-package org.seonhwan.android.veloginmobile
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import org.seonhwan.android.veloginmobile.databinding.ActivityLoginBinding
-
-class LoginActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityLoginBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityLoginBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        this.onClickLoginButton()
-    }
-
-    private fun onClickLoginButton(){
-        with(binding){
-            btLoginLogin.setOnClickListener {
-                val intent = Intent(this@LoginActivity, MainActivity::class.java)
-                startActivity(intent)
-                if(!isFinishing) finish()
-            }
-        }
-    }
-}
